feat: 점수 확인 기능 구현
