+++ src/main/java/blackjack/domain/Card.java
@@ -1,0 +1,24 @@
+package blackjack.domain;
+
+public class Card {
+
+    private final Number number;
+    private final Suit suit;
+
+    public Card(final Number number, final Suit suit) {
+        this.number = number;
+        this.suit = suit;
+    }
+
+    public int getScore() {
+        return number.getScore();
+    }
+
+    public String getNumberName() {
+        return number.getName();
+    }
+
+    public String getSuitName() {
+        return suit.getName();
+    }
+}
